Move HTML<->Lexical functions to @lexical/headless
