+++ packages/lexical-clipboard/src/clipboard.ts
@@ -20,8 +20,12 @@ import type {
   TextNode,
 } from 'lexical';
 
-import {$cloneWithProperties} from '@lexical/selection';
+import {
+  $cloneWithProperties,
+  $sliceSelectedTextNodeContent,
+} from '@lexical/selection';
 import {$findMatchingParent} from '@lexical/utils';
+import {$generateHtmlFromNodes, $generateNodesFromDOM} from '@lexical/headless';
 import {
   $createGridSelection,
   $createNodeFromParse,
@@ -41,8 +45,6 @@ import {
 } from 'lexical';
 import invariant from 'shared/invariant';
 
-const IGNORE_TAGS = new Set(['STYLE']);
-
 export function $getHtmlContent(editor: LexicalEditor): string | null {
   const selection = $getSelection();
 
@@ -58,77 +60,7 @@ export function $getHtmlContent(editor: LexicalEditor): string | null {
     return null;
   }
 
-  return $convertSelectedContentToHtml(editor, selection);
-}
-
-export function $appendSelectedNodesToHTML(
-  editor: LexicalEditor,
-  selection: RangeSelection | NodeSelection | GridSelection,
-  currentNode: LexicalNode,
-  parentElement: HTMLElement | DocumentFragment,
-): boolean {
-  let shouldInclude = currentNode.isSelected();
-  const shouldExclude =
-    $isElementNode(currentNode) && currentNode.excludeFromCopy('html');
-  let clone = $cloneWithProperties<LexicalNode>(currentNode);
-  clone = $isTextNode(clone) ? $splitClonedTextNode(selection, clone) : clone;
-  const children = $isElementNode(clone) ? clone.getChildren() : [];
-  const {element, after} = clone.exportDOM(editor);
-
-  if (!element) {
-    return false;
-  }
-
-  const fragment = new DocumentFragment();
-
-  for (let i = 0; i < children.length; i++) {
-    const childNode = children[i];
-    const shouldIncludeChild = $appendSelectedNodesToHTML(
-      editor,
-      selection,
-      childNode,
-      fragment,
-    );
-
-    if (
-      !shouldInclude &&
-      $isElementNode(currentNode) &&
-      shouldIncludeChild &&
-      currentNode.extractWithChild(childNode, selection, 'html')
-    ) {
-      shouldInclude = true;
-    }
-  }
-
-  if (shouldInclude && !shouldExclude) {
-    element.append(fragment);
-    parentElement.append(element);
-
-    if (after) {
-      const newElement = after.call(clone, element);
-      if (newElement) element.replaceWith(newElement);
-    }
-  } else {
-    parentElement.append(fragment);
-  }
-
-  return shouldInclude;
-}
-
-export function $convertSelectedContentToHtml(
-  editor: LexicalEditor,
-  selection: RangeSelection | NodeSelection | GridSelection,
-): string {
-  const container = document.createElement('div');
-  const root = $getRoot();
-  const topLevelChildren = root.getChildren();
-
-  for (let i = 0; i < topLevelChildren.length; i++) {
-    const topLevelNode = topLevelChildren[i];
-    $appendSelectedNodesToHTML(editor, selection, topLevelNode, container);
-  }
-
-  return container.innerHTML;
+  return $generateHtmlFromNodes(editor, selection);
 }
 
 export function $appendSelectedNodesToClone(
@@ -143,7 +75,9 @@ export function $appendSelectedNodesToClone(
   const shouldExclude =
     $isElementNode(currentNode) && currentNode.excludeFromCopy('clone');
   let clone = $cloneWithProperties<LexicalNode>(currentNode);
-  clone = $isTextNode(clone) ? $splitClonedTextNode(selection, clone) : clone;
+  clone = $isTextNode(clone)
+    ? $sliceSelectedTextNodeContent(selection, clone)
+    : clone;
   const children = $isElementNode(clone) ? clone.getChildren() : [];
   const nodeKeys = [];
   let shouldIncludeChildrenInRange = shouldIncludeInRange;
@@ -549,178 +483,4 @@ export function $generateNodes(nodeRange: {
   }
 
   return nodes;
-}
-
-function getConversionFunction(
-  domNode: Node,
-  editor: LexicalEditor,
-): DOMConversionFn | null {
-  const {nodeName} = domNode;
-
-  const cachedConversions = editor._htmlConversions.get(nodeName.toLowerCase());
-
-  let currentConversion: DOMConversion | null = null;
-
-  if (cachedConversions !== undefined) {
-    cachedConversions.forEach((cachedConversion) => {
-      const domConversion = cachedConversion(domNode);
-
-      if (domConversion !== null) {
-        if (
-          currentConversion === null ||
-          currentConversion.priority < domConversion.priority
-        ) {
-          currentConversion = domConversion;
-        }
-      }
-    });
-  }
-
-  return currentConversion !== null ? currentConversion.conversion : null;
-}
-
-function $createNodesFromDOM(
-  node: Node,
-  editor: LexicalEditor,
-  forChildMap: Map<string, DOMChildConversion> = new Map(),
-  parentLexicalNode?: LexicalNode | null | undefined,
-): Array<LexicalNode> {
-  let lexicalNodes: Array<LexicalNode> = [];
-
-  if (IGNORE_TAGS.has(node.nodeName)) {
-    return lexicalNodes;
-  }
-
-  let currentLexicalNode = null;
-  const transformFunction = getConversionFunction(node, editor);
-  const transformOutput = transformFunction ? transformFunction(node) : null;
-  let postTransform = null;
-
-  if (transformOutput !== null) {
-    postTransform = transformOutput.after;
-    currentLexicalNode = transformOutput.node;
-
-    if (currentLexicalNode !== null) {
-      for (const [, forChildFunction] of forChildMap) {
-        currentLexicalNode = forChildFunction(
-          currentLexicalNode,
-          parentLexicalNode,
-        );
-
-        if (!currentLexicalNode) {
-          break;
-        }
-      }
-
-      if (currentLexicalNode) {
-        lexicalNodes.push(currentLexicalNode);
-      }
-    }
-
-    if (transformOutput.forChild != null) {
-      forChildMap.set(node.nodeName, transformOutput.forChild);
-    }
-  }
-
-  // If the DOM node doesn't have a transformer, we don't know what
-  // to do with it but we still need to process any childNodes.
-  const children = node.childNodes;
-  let childLexicalNodes = [];
-
-  for (let i = 0; i < children.length; i++) {
-    childLexicalNodes.push(
-      ...$createNodesFromDOM(
-        children[i],
-        editor,
-        forChildMap,
-        currentLexicalNode,
-      ),
-    );
-  }
-
-  if (postTransform != null) {
-    childLexicalNodes = postTransform(childLexicalNodes);
-  }
-
-  if (currentLexicalNode == null) {
-    // If it hasn't been converted to a LexicalNode, we hoist its children
-    // up to the same level as it.
-    lexicalNodes = lexicalNodes.concat(childLexicalNodes);
-  } else {
-    if ($isElementNode(currentLexicalNode)) {
-      // If the current node is a ElementNode after conversion,
-      // we can append all the children to it.
-      currentLexicalNode.append(...childLexicalNodes);
-    }
-  }
-
-  return lexicalNodes;
-}
-
-export function $generateNodesFromDOM(
-  dom: Document,
-  editor: LexicalEditor,
-): Array<LexicalNode> {
-  let lexicalNodes = [];
-  const elements: Array<Node> = dom.body ? Array.from(dom.body.childNodes) : [];
-  const elementsLength = elements.length;
-
-  for (let i = 0; i < elementsLength; i++) {
-    const element = elements[i];
-
-    if (!IGNORE_TAGS.has(element.nodeName)) {
-      const lexicalNode = $createNodesFromDOM(element, editor);
-
-      if (lexicalNode !== null) {
-        lexicalNodes = lexicalNodes.concat(lexicalNode);
-      }
-    }
-  }
-
-  return lexicalNodes;
-}
-
-export function $splitClonedTextNode(
-  selection: RangeSelection | GridSelection | NodeSelection,
-  clone: TextNode,
-): LexicalNode {
-  if (
-    clone.isSelected() &&
-    !clone.isSegmented() &&
-    !clone.isToken() &&
-    ($isRangeSelection(selection) || $isGridSelection(selection))
-  ) {
-    const anchorNode = selection.anchor.getNode();
-    const focusNode = selection.focus.getNode();
-    const isAnchor = clone.is(anchorNode);
-    const isFocus = clone.is(focusNode);
-
-    if (isAnchor || isFocus) {
-      const isBackward = selection.isBackward();
-      const [anchorOffset, focusOffset] = selection.getCharacterOffsets();
-      const isSame = anchorNode.is(focusNode);
-      const isFirst = clone.is(isBackward ? focusNode : anchorNode);
-      const isLast = clone.is(isBackward ? anchorNode : focusNode);
-      let startOffset = 0;
-      let endOffset = undefined;
-
-      if (isSame) {
-        startOffset = anchorOffset > focusOffset ? focusOffset : anchorOffset;
-        endOffset = anchorOffset > focusOffset ? anchorOffset : focusOffset;
-      } else if (isFirst) {
-        const offset = isBackward ? focusOffset : anchorOffset;
-        startOffset = offset;
-        endOffset = undefined;
-      } else if (isLast) {
-        const offset = isBackward ? anchorOffset : focusOffset;
-        startOffset = 0;
-        endOffset = offset;
-      }
-
-      clone.__text = clone.__text.slice(startOffset, endOffset);
-      return clone;
-    }
-  }
-
-  return clone;
 }
