+++ packages/lexical-headless/LexicalHeadless.d.ts
@@ -11,6 +11,9 @@ import type {
   LexicalNode,
   EditorState,
   EditorThemeClasses,
+  RangeSelection,
+  NodeSelection,
+  GridSelection,
 } from 'lexical';
 
 import {Class} from 'utility-types';
@@ -25,3 +28,13 @@ export function createHeadlessEditor(editorConfig?: {
   disableEvents?: boolean;
   readOnly?: boolean;
 }): LexicalEditor;
+
+export function $generateHtmlFromNodes(
+  editor: LexicalEditor,
+  selection?: RangeSelection | NodeSelection | GridSelection | null,
+): string;
+
+export function $generateNodesFromDOM(
+  dom: Document,
+  editor: LexicalEditor,
+): Array<LexicalNode>;
