+++ packages/lexical-playground/vite.config.js
@@ -32,6 +32,10 @@ const moduleResolution = [
     replacement: path.resolve('../lexical-text/src/index.ts'),
   },
   {
+    find: '@lexical/headless',
+    replacement: path.resolve('../lexical-headless/src/index.ts'),
+  },
+  {
     find: '@lexical/hashtag',
     replacement: path.resolve('../lexical-hashtag/src/index.ts'),
   },
