+++ tsconfig.json
@@ -23,6 +23,7 @@
       "@lexical/text": ["./packages/lexical-text/LexicalText.d.ts"],
       "@lexical/offset": ["./packages/lexical-offset/LexicalOffset.d.ts"],
       "@lexical/history": ["./packages/lexical-history/LexicalHistory.d.ts"],
+      "@lexical/headless": ["./packages/lexical-headless/LexicalHeadless.d.ts"],
       "@lexical/utils": ["./packages/lexical-utils/LexicalUtils.d.ts"],
       "@lexical/code": ["./packages/lexical-code/LexicalCode.d.ts"],
       "@lexical/plain-text": [
